+++ custom_components/solarcharger/datetime.py
@@ -148,7 +148,9 @@ async def async_setup_entry(
                 )
 
         if len(datetimes) > 0:
-            coordinator.charge_controls[subentry.subentry_id].datetimes = datetimes
+            coordinator.device_controls[
+                subentry.subentry_id
+            ].controller.charge_control.datetimes = datetimes
             async_add_entities(
                 datetimes.values(),
                 update_before_add=False,