+++ custom_components/solarcharger/input_datetime.py
@@ -224,7 +224,9 @@ async def async_setup_entry(
                 )
 
         if len(input_times) > 0:
-            coordinator.charge_controls[subentry.subentry_id].input_times = input_times
+            coordinator.device_controls[
+                subentry.subentry_id
+            ].controller.charge_control.input_times = input_times
             async_add_entities(
                 input_times.values(),
                 update_before_add=False,