+++ custom_components/solarcharger/number.py
@@ -504,7 +504,9 @@ async def async_setup_entry(
                 )
 
         if len(numbers) > 0:
-            coordinator.charge_controls[subentry.subentry_id].numbers = numbers
+            coordinator.device_controls[
+                subentry.subentry_id
+            ].controller.charge_control.numbers = numbers
             async_add_entities(
                 numbers.values(),
                 update_before_add=False,