+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v0.3.9 2026-02-08
+- Refactored charger controller.
 - Removed redundant start charge button.
 
 ## v0.3.8.1 2026-02-07